pyproject.toml
[tool.poetry]
name = "batchgrader"
version = "0.6.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "batchgrader", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
annotated-types = "^0.7.0"
anyio = "^4.9.0"
certifi = ">=2025.4.26"
charset-normalizer = ">=3.4.2"
distro = ">=1.9.0"
h11 = ">=0.16.0"
httpcore = ">=1.0.9"
httpx = ">=0.28.1"
idna = ">=3.10"
iniconfig = ">=2.1.0"
jinja2 = ">=3.1.6"
jiter = ">=0.9.0"
markupsafe = ">=3.0.2"
numpy = ">=2.2.5"
openai = ">=1.78.0"
packaging = ">=25.0"
pandas = ">=2.2.3"
pluggy = ">=1.6.0"
pydantic = ">=2.11.4"
pydantic-core = ">=2.33.2"
pytest = ">=8.3.5"
pytest-asyncio = ">=0.26.0"
pytest-mock = ">=3.14.0"
python-dateutil = ">=2.9.0"
python-dotenv = ">=1.1.0"
pytz = ">=2025.2"
pyyaml = ">=6.0.2"
regex = ">=2024.11.6"
requests = ">=2.32.3"
rich = ">=14.0.0"
six = ">=1.17.0"
sniffio = ">=1.3.1"
tenacity = "^8.2.3"
tiktoken = ">=0.9.0"
tqdm = ">=4.67.1"
typing-extensions = ">=4.13.2"
typing-inspection = ">=0.4.0"
tzdata = ">=2025.2"
urllib3 = ">=2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
